aider: Refactor `fetchData` function in `apolloClient.ts` to explicitly define the type of the `query` parameter as `DocumentNode`.
# Aider chat conversation:

USER: in `ApolloClient` i got a error in line 8 El parámetro 'query' tiene un tipo 'any' implícitamente.ts(7006)

ASSISTANT: The error message you're seeing is because TypeScript is not able to infer the type of the `query` parameter in the `fetchData` function. To fix this, we can explicitly define the type of the `query` parameter.

The `query` parameter should be of type `DocumentNode`, which is the type of object that the `gql` function returns. We can import `DocumentNode` from the `@apollo/client` package and use it to type the `query` parameter.

Here is the edit block for this change:

```ts
lib/apolloClient.ts
<<<<<<< HEAD
export async function fetchData(query) {
=======
import { DocumentNode } from '@apollo/client';

export async function fetchData(query: DocumentNode) {
>>>>>>> updated
  const { data } = await client.query({
    query
  });

  return data;
}
```