aider: Convert `launch_date_local` to a formatted date string in the "Wesnesday 03/06/2020, 21:25:00" format in the `LaunchItem` component.
# Aider chat conversation:

USER: if i got "2020-06-03T21:25:00-04:00" i want to see for example this format: Wesnesday 03/06/2020, 21:25:00

ASSISTANT: To display the date in the format you specified, you can use the `toLocaleDateString` and `toLocaleTimeString` methods of the JavaScript `Date` object. The `toLocaleDateString` method can be used to format the date part and the `toLocaleTimeString` method can be used to format the time part.

Here's the change you need to make:

<source>typescript
components/LaunchItem.tsx
<<<<<<< HEAD
      <p><strong>Launch Date:</strong> { new Date(launch.launch_date_local).toLocaleDateString("es-ES", {timeZone: 'UTC'}) }</p>
=======
      <p><strong>Launch Date:</strong> { new Date(launch.launch_date_local).toLocaleDateString("es-ES", {weekday: 'long', day: '2-digit', month: '2-digit', year: 'numeric', timeZone: 'UTC'}) }, { new Date(launch.launch_date_local).toLocaleTimeString("es-ES", {hour: '2-digit', minute: '2-digit', second: '2-digit', timeZone: 'UTC'}) }</p>
>>>>>>> updated
</source>

This change will ensure that the `launch_date_local` is correctly converted to a date string in the format "Wesnesday 03/06/2020, 21:25:00" in the `LaunchItem` component.